chore: re-organize fragments indexed nfa view
