+++ ui/src/views/indexed-nfa/indexed-nfa.styles.ts
@@ -56,7 +56,6 @@ export const IndexedNFAStyles = {
         lineHeight: 1.35,
         fontWeight: 700,
 
-        // maxWidth: '10rem',
         overflow: 'hidden',
         textOverflow: 'ellipsis',
         whiteSpace: 'nowrap',
@@ -151,12 +150,6 @@ export const IndexedNFAStyles = {
       lineHeight: 1.35,
       fontWeight: 700,
     }),
-    SectionHeading: styled('h2', {
-      fontSize: '$xl',
-      lineHeight: 1.2,
-      fontWeight: 700,
-      marginTop: Spacing,
-    }),
     AccessPoint: {
       List: styled('div', {
         display: 'flex',
@@ -186,6 +179,12 @@ export const IndexedNFAStyles = {
         color: '$slate12',
         fontSize: '$lg',
       }),
+      NoResults: styled('div', {
+        display: 'flex',
+        justifyContent: 'center',
+
+        fontSize: '$lg',
+      }),
     },
     Divider: {
       Line: styled('span', {
@@ -197,144 +196,6 @@ export const IndexedNFAStyles = {
         height: '0.375rem',
         backgroundColor: '$slate11',
         borderRadius: '100%',
-      }),
-    },
-    Paragraph: styled('p', {
-      color: '$slate11',
-      lineHeight: 1.43,
-    }),
-    DataContainer: styled('div', {
-      display: 'flex',
-      flexDirection: 'column',
-      border: '1px solid $slate6',
-      borderRadius: '$lg',
-      padding: Spacing,
-      gap: `$1`,
-    }),
-    DataList: styled('div', {
-      display: 'flex',
-      flexWrap: 'wrap',
-      gap: '$5',
-    }),
-    VerificationBanner: styled('div', {
-      position: 'relative',
-      display: 'flex',
-      alignItems: 'center',
-      border: '1px solid $slate6',
-      borderRadius: '$lg',
-      padding: '$8 $5',
-      fontWeight: 700,
-      overflow: 'hidden',
-
-      '&:after': {
-        content: '""',
-        position: 'absolute',
-        right: '-$5',
-        top: '-$10',
-        bottom: '-$10',
-        left: '84%',
-        borderRadius: '80% 0 0 80%',
-      },
-
-      variants: {
-        verified: {
-          true: {
-            borderColor: '$green11',
-            color: '$green11',
-            '&:after': {
-              backgroundColor: '$green11',
-            },
-          },
-          false: {
-            borderColor: '$red11',
-            color: '$red11',
-            '&:after': {
-              backgroundColor: '$red11',
-            },
-          },
-        },
-      },
-    }),
-    Table: {
-      Container: styled('div', {
-        border: '1px solid $slate6',
-        borderRadius: '10px',
-        padding: '0 $5',
-
-        maxHeight: '15.125rem',
-        overflow: 'auto',
-      }),
-      Root: styled('table', {
-        width: 'calc(100% + 2 * $space$5)',
-        margin: '0 -$5',
-      }),
-      Head: styled('thead', {
-        position: 'sticky',
-        top: 0,
-        backgroundColor: '$black',
-
-        '&:after': {
-          position: 'absolute',
-          content: '""',
-          bottom: 0,
-          left: 0,
-          right: 0,
-          borderBottom: '1px solid $slate6',
-        },
-      }),
-      Row: styled('tr'),
-      Data: styled('td', {
-        padding: '$3',
-        maxWidth: '10rem',
-        overflow: 'hidden',
-        textOverflow: 'ellipsis',
-        whiteSpace: 'nowrap',
-      }),
-      Body: styled('tbody', {
-        tr: {
-          '&:hover': {
-            backgroundColor: '$slate6',
-            cursor: 'pointer',
-          },
-        },
-      }),
-      Marker: styled('span', {
-        display: 'block',
-        margin: 'auto',
-        width: '0.5625rem',
-        height: '0.5625rem',
-        borderRadius: '$full',
-        backgroundColor: '$slate6',
-
-        variants: {
-          variant: {
-            active: {
-              backgroundColor: '$green11',
-            },
-            inactive: {
-              backgroundColor: '$slate8',
-            },
-          },
-          text: {
-            true: {
-              fontSize: '$xs',
-              padding: '0 $2',
-              width: 'fit-content',
-              height: 'fit-content',
-            },
-          },
-        },
-
-        compoundVariants: [
-          {
-            variant: 'active',
-            text: true,
-            css: {
-              color: '$green11',
-              backgroundColor: '$green3',
-            },
-          },
-        ],
       }),
     },
   },
