+++ ui/src/views/indexed-nfa/tabs/index.ts
