+++ ui/src/views/indexed-nfa/tabs/tabs.styles.ts
