+++ ui/src/views/indexed-nfa/tabs/tabs.tsx
